+++ server/app/models/statisticwh.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, DateTime
+from sqlalchemy import Column, ForeignKey, Double, String, DateTime
 from app.utils.db import Base
 from sqlalchemy.orm import mapped_column, relationship, Mapped
 from datetime import datetime
@@ -13,7 +13,7 @@ class StatisticWH(Base):
         primary_key=True, autoincrement=True)
 
     Date: Mapped[datetime] = mapped_column(DateTime, default=datetime.now)
-    Height: Mapped[int] = mapped_column(Integer)
-    Weight: Mapped[int] = mapped_column(Integer)
+    Height: Mapped[Double] = mapped_column(Double)
+    Weight: Mapped[Double] = mapped_column(Double)
 
     UserID: Mapped[int] = mapped_column(ForeignKey("user.UserID"))
