+++ server/app/schemas/food.py
@@ -4,10 +4,10 @@ from typing import Any
 class FoodBase(BaseModel):
     FoodID: int
     Name: str
-    Calories: int
-    Protein: int
-    Fats: int
-    Carbonates: int 
+    Calories: float
+    Protein: float
+    Fats: float
+    Carbonates: float 
 
 class FoodDTO(FoodBase):
     Category: str
