+++ server/app/schemas/userstatistic.py
@@ -7,12 +7,12 @@ class UserStatisticBase(BaseModel):
 
 class UserStatisticDTO(UserStatisticBase):
     Date: datetime 
-    UserHeight: int 
-    UserWeight: int 
-    Calories: int 
-    Protein: int 
-    Fats: int 
-    Carbonates: int 
+    UserHeight: float 
+    UserWeight: float 
+    Calories: float 
+    Protein: float 
+    Fats: float 
+    Carbonates: float 
 
     class Config:
         from_attributes = True
